Rename the "VerificationStatus" folder to "AppSwitches" and change all the imports in the "hooks.js", "Login.js", "Signup.js" and "Store.js" files leading to the files located in this folder.
