+++ src/store/AppSwitches/Reducer.js
