+++ src/store/AppSwitches/Selectors.js
